requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
click==8.1.3
colorama==0.4.5
databases==0.6.1
dj-config-url==0.1.1
dj-database-url==1.0.0
Django==4.1.1
django-rest-framework==0.1.0
django-scheduler==0.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.0
greenlet==1.1.3
icalendar==4.1.0
inflection==0.5.1
jsonschema==4.16.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
